--- a/bstream.go
+++ b/bstream.go
@@ -100,11 +100,11 @@ func (b *bstream) readBit() (bit, error) {
 	}
 
 	if b.count == 0 {
-		b.stream = b.stream[1:]
 		// did we just run out of stuff to read?
-		if len(b.stream) == 0 {
+		if len(b.stream) == 1 {
 			return false, io.EOF
 		}
+		b.stream = b.stream[1:]
 		b.count = 8
 	}