@dgryski can you give these a critical look and let me know whether they look good or are stupid, and if so, why. thanks :)