@dgryski 2016-02-23
I'll run some benchmarks today and see if they change