@dgryski 2016-02-23
Ran 30 interations of my benchmarks and stuffed them into https://godoc.org/rsc.io/benchstat

Looks like no significant changes.  If anything, Encode got a teeny tiny bit slower.

```
<dgryski@kamek[go-tsz] \ʕ◔ϖ◔ʔ/ > benchstat bench.old bench.new 
name      old time/op  new time/op  delta
Encode-4  19.1µs ± 2%  19.4µs ± 3%  +1.64%  (p=0.000 n=29+30)
Decode-4  7.26µs ± 2%  7.28µs ± 2%    ~     (p=0.534 n=29+30)
```