@dgryski 2016-02-23
The test data set is two hours of secondly data from a production system.