@dgryski 2016-02-23
Oops, that was the regular two hours of minutely data.  Reran again with the two hours of secondly data.  Still no change:
```
<dgryski@kamek[go-tsz] \ʕ◔ϖ◔ʔ/ > benchstat bench.old bench.new 
name       old time/op  new time/op  delta
EncodeB-4  1.31ms ± 2%  1.32ms ± 4%  +1.33%  (p=0.000 n=29+30)
DecodeB-4   595µs ± 2%   599µs ± 4%    ~     (p=0.130 n=30+30)
Encode-4   19.3µs ± 2%  19.4µs ± 3%    ~     (p=0.120 n=28+30)
Decode-4   7.28µs ± 2%  7.31µs ± 2%    ~     (p=0.085 n=27+29)
```