@Dieterbe 2016-02-23
Ok but benchmarks aside, what do you think of the commits, logically speaking? Especially the first and third make sense to me, I'm not sure about the second. @dgryski