@dgryski 2016-02-23
The first commit makes sense. The second one I agree there are too many variables, and as was shown by the benchmarks it didn't make a difference.  The third one I think makes for slightly less clear code.  I like the fact that we return EOF when  `len(b.stream)==0`.  We've totally consumed `b.stream` and the slice is now empty.  Personal preference, I guess.  (Much like I dislike `unless` in perl and try to always write numerical comparisons with `<` instead of a mixture of `<` and `>`, etc...)