@dgryski 2016-02-23
You can merge https://github.com/dgryski/go-tsz/commit/c52a434821dd04a69468fcb9a49501c8927be001 if you want.